feat(extension.js): add support for Twig language in languageSupport array
The change was made to include 'twig' in the `lenguageSupport` array in the `extension.js` file. This allows the extension to provide support for the Twig language. The addition of Twig language support enhances the functionality of the extension by enabling syntax highlighting and other language-specific features for Twig files.